scripts/docker/requirements.txt
biopython==1.81
orfipy==0.0.4
scikit-learn==1.3.2
pandas==2.0.3
gradio
transformers
trl 
peft